// lbo-sdk/src/plugin/cms/models/TileWidgetTypeEnum.ts
export enum TileWidgetTypeEnum {
    EMPTY = '',
    BANNER = "banner",
    GRID = "grid",
    TILES = "tiles",
    TOP_GAMES = "topGames",
    RECENTLY_PLAYED_GAMES = "recentlyPlayedGames",
    ATOZ = "atoz",
    RECOMMENDED_GAMES = "recommendedGames",
    TAGGED_GAMES = "taggedGames",
    JACKPOT_GRID = "jackpotGrid",
    JACKPOT_TILE = "jackpotTile",
    DFG = "dfg"
}